Don't delete Vprj file if VRD encode job follows
git-svn-id: https://svn.code.sf.net/p/kmttg/code/trunk@4049 b0df3fdd-2057-4f73-b226-dfc9190e8f7a

+++ src/com/tivo/kmttg/task/vrdreview.java
@@ -218,8 +218,14 @@ public class vrdreview extends baseTask implements Serializable {
             
             if (config.VrdReview_noCuts == 1) {
                if (config.RemoveComcutFiles == 1) {
-                  if (file.delete(job.vprjFile))
-                     log.print("(Deleted vprj file: " + job.vprjFile + ")");
+                  if (jobMonitor.getJobInFamily(job, "vrdencode") == null) {
+                     // No VRD encode task following, so OK to delete .Vprj file
+                     if (file.delete(job.vprjFile))
+                        log.print("(Deleted vprj file: " + job.vprjFile + ")");
+                  } else {
+                     // VRD encode task follows, so may need to use .Vprj file
+                     log.warn("Not deleting Vprj file, since VRD encode task follows");
+                  }
                   String xclFile = job.mpegFile + ".Xcl";
                   if (file.isFile(xclFile) && file.delete(xclFile))
                      log.print("(Deleted xcl file: " + xclFile + ")");
